requirements.txt
streamlit
langchain
chromadb
huggingface-hub
transformers
requests
python-dotenv
urllib3